Struct power_system::application::Application

pub struct Application {
    adc: ADC<'static, ADC1>,
    uart: Uart<'static, UART1>,
    onboard_led: OnboardLed,
    battery_measurement_pin: BatteryMeasurementPin,
}

Fields
adc: ADC<'static, ADC1>uart: Uart<'static, UART1>onboard_led: OnboardLedbattery_measurement_pin: BatteryMeasurementPin
Implementations

impl Application

pub fn init(peripherals: Peripherals) -> Self

Initialises all the peripherals which the Application will use.

pub fn run(self, executor: &'static mut Executor) -> !

Runs the application by spawning each of the Application’s tasks